// Client/src/app/_models/ContractDto.ts
export class ContractDto {
    id:string;
    employeeId: string;
    expiredDate: Date;
    contractDate: Date;
    contracType: number;
    contractPerson: String;
    contracPerson: String;
    remark:string;
    createdUser:number;
    createdDate:Date;
    updatedUser:string;
    updatedDate:Date;
   
}